fix(consultas e exercicios): corrigir header e exercicios
