+++ src/@core/services/plan/professional/ExercisesSelectionHandlerService.ts
