+++ src/app/clients/appointments/page.tsx
@@ -45,7 +45,7 @@ export default function AppointmentsScreen() {
   return (
     <ThemeProvider theme={theme}>
       <NextUIProvider className="default">
-        <AppHeader />
+        <AppHeader isProfessional />
 
         <AppContainer style={{ justifyContent: 'start' }}>
           <AppButtonLinkRectOutline
