+++ src/app/clients/page.tsx
@@ -14,7 +14,7 @@ export default function ClientsScreen() {
   return (
     <ThemeProvider theme={theme}>
       <NextUIProvider className="default">
-        <AppHeader />
+        <AppHeader isProfessional />
 
         <AppContainer style={{ justifyContent: 'start' }}>
           <AppDashboardHeader
