+++ src/app/clients/plan/edit/[planId]/page.tsx
@@ -4,8 +4,8 @@ import { useMemo } from 'react'
 import 'react-toastify/dist/ReactToastify.css'
 
 import { PlanBuildStageEnum } from '@components/clients/professionals/sharedProps/PlanBuilderStage'
-import PlanBuilderScreen from '../../new/page'
 import { GetRandomTrainingPlanService } from '@core/services/seed/trainingPlan/GetRandomTrainingPlanService'
+import { AppProfessionalPlanManager } from '@components/clients/professionals/planBuilder/AppProfessionalPlanManager'
 
 interface EditPlanBuilderScreenProps {
   params: {
@@ -20,7 +20,7 @@ export default function EditPlanBuilderScreen(
   const trainingPlan = useMemo(() => GetRandomTrainingPlanService.exec(), [])
 
   return (
-    <PlanBuilderScreen
+    <AppProfessionalPlanManager
       initialStage={PlanBuildStageEnum.CONFIRM}
       preLoadedTrainingPlan={trainingPlan}
       skipClientSelection
