+++ src/components/exercises/AppProgressExercises.tsx
@@ -8,6 +8,9 @@ import { ProgressExercisesStyle } from './ProgressExercisesStyle'
 import { AppCircularProgress } from '@components/common/Progress/AppCircularProgress'
 import Image from 'next/image'
 import Link from 'next/link'
+import { ExerciseRepo } from '@core/repositories/ExerciseRepo'
+
+const exercises = ExerciseRepo.getAllExercises()
 
 export function AppProgressExercises() {
   return (
@@ -39,10 +42,7 @@ export function AppProgressExercises() {
               width={100}
               height={100}
             />
-            <Link
-              href="/exercises/alongamento/exercício_da_cadeira"
-              id="next-exercise-btn"
-            >
+            <Link href={exercises[0].href!} id="next-exercise-btn">
               <h3>Próximo</h3>
               <small>Fácil</small>
             </Link>
